Move parse utils in their own module
