+++ crates/ark/src/interface.rs
@@ -63,7 +63,6 @@ use harp::environment::R_ENVS;
 use harp::exec::r_check_stack;
 use harp::exec::r_peek_error_buffer;
 use harp::exec::r_sandbox;
-use harp::exec::r_source;
 use harp::exec::RFunction;
 use harp::exec::RFunctionExt;
 use harp::library::RLibraries;
@@ -71,6 +70,7 @@ use harp::line_ending::convert_line_endings;
 use harp::line_ending::LineEnding;
 use harp::object::r_null_or_try_into;
 use harp::object::RObject;
+use harp::r_source;
 use harp::r_symbol;
 use harp::routines::r_register_routines;
 use harp::session::r_traceback;
