+++ crates/ark/src/modules.rs
@@ -9,10 +9,10 @@ use anyhow::anyhow;
 use harp::environment::Environment;
 use harp::environment::R_ENVS;
 use harp::eval::r_parse_eval;
-use harp::exec::r_parse_exprs_with_srcrefs;
-use harp::exec::r_source_str_in;
 use harp::exec::RFunction;
 use harp::exec::RFunctionExt;
+use harp::r_parse_exprs_with_srcrefs;
+use harp::r_source_str_in;
 use harp::r_symbol;
 use harp::utils::r_poke_option;
 use libr::Rf_ScalarLogical;
