+++ crates/ark/src/shell.rs
@@ -38,11 +38,11 @@ use crossbeam::channel::unbounded;
 use crossbeam::channel::Receiver;
 use crossbeam::channel::Sender;
 use harp::environment::R_ENVS;
-use harp::exec::r_parse_vector;
-use harp::exec::ParseResult;
 use harp::line_ending::convert_line_endings;
 use harp::line_ending::LineEnding;
 use harp::object::RObject;
+use harp::r_parse_vector;
+use harp::ParseResult;
 use log::*;
 use serde_json::json;
 use stdext::spawn;
